Let the finance generator own the surface it already generates
It built the whole live admin and public surface and then copied three
hard-coded travel-credit paths out of it, leaving 99 of 102 admin paths and 11
of 12 public paths carried through from the committed artifact untouched.
`verify:openapi-drift` reported both documents as matching their routes.

They did not. Regenerating what was already being generated found:

  - `/v1/admin/finance/invoices/{id}/external-refs/{refId}/supersede` missing
    from the document entirely
  - `providerConnectionId` absent from 18 operations that return it
  - `targetType` enums missing `booking_session` in 14 operations, so the
    contract declared a value the API actually returns to be impossible
  - 22 differing `required` lists, and three profitability report fields
  - 148 operations with no `x-voyant-api-id` stamp

Two things had to change beyond widening the loop.

Sibling documents own some of that surface. `tax-settings`, `tax-preview` and
`payment-link` are their own artifacts with their own graph API bundles, so
claiming everything the surface produces documented `tax-settings` twice and
`verify:deployment-graph-openapi-coverage` rejected it as a duplicate authority.
The exclusion is read from the package's tracked artifacts rather than listed in
the script, so a document that takes over a path stops this one claiming it with
nobody editing a list.

And `paths` is rebuilt wholesale rather than merged into. Merging only ever
adds, so a path the surface stopped serving survived forever with nothing
regenerating or comparing it — the exact defect being closed here, and it would
have left a residue of its own. `legacyPaths` needs no special case now.

Regenerating through the 3.1 builder also cleared the three stale `nullable`
keywords in the public document, so that baseline drops from 54 to 51.

The script is renamed off `generate-travel-credit-openapi.ts`; it stopped being
about travel credits.
