+++ scripts/agent-write-coverage-baseline.json
@@ -10,7 +10,7 @@
   "@voyant-travel/cruises": 21,
   "@voyant-travel/custom-fields": 2,
   "@voyant-travel/distribution": 51,
-  "@voyant-travel/finance": 44,
+  "@voyant-travel/finance": 45,
   "@voyant-travel/flights": 3,
   "@voyant-travel/inventory": 42,
   "@voyant-travel/legal": 16,
