+++ scripts/checks/openapi/path-ownership-baseline.json
@@ -11,8 +11,6 @@
     "document so each half is owned by the route module that produces it."
   ],
   "unownedPaths": {
-    "packages/finance/openapi/admin/finance.json": 99,
-    "packages/finance/openapi/public-api/finance.json": 11,
     "packages/legal/openapi/admin/legal.json": 15,
     "packages/public-api/openapi/public-api/public-api.json": 12
   }
